pyproject.toml
[project]
name = "djangosoftdesk"
version = "0.1.0"
description = "OpenClassRooms Project"
authors = [
    {name = "Igor",email = "pierre.zarebski@live.fr"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.15.0,<4.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"